requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
sqlalchemy
alembic
psycopg2-binary
redis
httpx
python-multipart
python-jose[cryptography]
passlib[bcrypt]
anthropic
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
gspread
python-dotenv
celery
PyYAML
asyncpg
aioredis
python-dateutil
pytz
requests
google-cloud-speech
google-cloud-texttospeech
fastapi
uvicorn
pydantic
pydantic-settings
sqlalchemy
alembic
psycopg2-binary
redis
httpx
python-multipart
python-jose[cryptography]
passlib[bcrypt]
anthropic
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
gspread
python-dotenv
celery
PyYAML
asyncpg
aioredis
python-dateutil
pytz
requests
aiogram
